requirements.txt
# requirements.txt
streamlit>=1.24.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
yfinance>=0.2.30
pandas>=2.0.0
newsapi-python>=0.2.7
requests>=2.31.0
plotly>=5.18.0
numpy>=1.24.0
python-dateutil>=2.8.2